services/payment-service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka>=2.13.2",
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "gunicorn>=25.1.0",
    "kafka-python>=2.3.0",
    "stripe>=14.4.1",
]
